Move faker keywords to a mixin class so that they can be used both with selenium and playwright
